readme, refactoring
