+++ .gitignore
@@ -1,7 +1,9 @@
 .DS_Store
 node_modules
 assets
-# /dist
+dist
+vuetify
+functions
 
 # local env files
 .env.local
@@ -9,8 +11,6 @@ assets
 
 # Log files
 npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
 
 # Editor directories and files
 .idea
