+++ README.md
@@ -1,29 +1,27 @@
-# test
+## Project description
+```
+Service to assist in social cooperations, including such possibilities as:
+ - Creating online project presentation with donations demands(Description, preliminary outlay, photos, form to get  participants'/donors' contacts)
+ - Public list of donations and participants/donors for every project
+ - Managing donations in User Page by user-project-creator
+ - Editing projects by user-project-creator
+ - Adding final outlay and final photos after project realization by user-project-creator
+```
+### Stack in use
+```
+Vue.js
+Vuex
+Vuetify(components)
+Firebase(authentication, database, functions)
+Sass
+Git(+GitHub)
+```
+### Functionality realized
+```
+Creating and editing projects by user via database Firebase,
+Users authentication with Firebase,
+User Page,
+Internationalization using i18n (English + Russian),
+Autoresizing uploaded images
+```
 
-## Project setup
-```
-npm install
-```
-
-### Compiles and hot-reloads for development
-```
-npm run serve
-```
-
-### Compiles and minifies for production
-```
-npm run build
-```
-
-### Run your tests
-```
-npm run test
-```
-
-### Lints and fixes files
-```
-npm run lint
-```
-
-### Customize configuration
-See [Configuration Reference](https://cli.vuejs.org/config/).
