+++ firebase.json
@@ -8,3 +8,15 @@
       ]
   }
 }
+
+{
+  "hosting": {
+    "public": "dist",
+    "rewrites": [
+      {
+        "source": "**",
+        "destination": "/index.html"
+      }
+    ]
+  }
+}
