+++ public/index.html
@@ -17,17 +17,12 @@
       box-sizing: border-box;
     }
     .header {
-      /*display: block;*/
       display: flex;
-      /*position: absolute;*/
       position: fixed;
-      /*flex-wrap: wrap;*/
       top: 0;
       z-index: 6;
       width: 100%;
-      /*max-width: 1000px;*/
       margin: auto;
-      /*justify-content: center;*/
       height: 50px;
       background-color: #fff;
       border-bottom: 1px solid #1B5E20;
@@ -96,13 +91,9 @@
         top: 50px;
         left: -260px;
         width: 250px;
-        /*height: 500px;*/
         z-index: 1;
         transition: 500ms ease-in-out;
-        /*transform: translateX(-110%);*/
         opacity: 0;
-        /*margin-left: -15px;*/
-        /*margin-top: 1px;*/
         padding: 15px 0;
      }
      .side-logo {
@@ -147,7 +138,7 @@
         letter-spacing: 1px;
         text-align: center;
         background-color: rgba(0, 0, 0, 0.5);
-        transform: translateY(400px);
+        transform: translateY(50vh);
       }
       @media (max-width: 840px){
         .features__title {
