+++ src/components/footer/Footer.vue
@@ -20,5 +20,5 @@ export default {
 };
 </script>
 <style lang="sass">
-  @import "./../../sass/_footer.sass"
+  @import "@/sass/_footer.sass"
 </style>
