+++ src/components/header/Header.vue
@@ -95,7 +95,7 @@ export default {
       selected: 'English',
       languages: ['English','Russian'],
       locale: 'en',
-      userIsCreator: false
+      // userIsCreator: false
     }
   },
   computed: {
@@ -104,6 +104,9 @@ export default {
     },
     loading(){
       return this.$store.getters.loading
+    },
+    projects(){
+      return this.$store.getters.projects
     }
   },
   methods: {
@@ -124,9 +127,10 @@ export default {
       this.$store.dispatch('initOutlay', this.locale)
     },
     showMenu() {
-      this.userIsCreator = this.$store.getters.projects.find((project) =>{
-        return project.creatorId == this.$store.getters.userId
-      })
+
+      // this.userIsCreator = this.$store.getters.projects.find((project) =>{
+      //   return project.creatorId == this.$store.getters.userId
+      // })
       this.isOpen = !this.isOpen
       this.isActive = !this.isActive
     },
@@ -143,5 +147,5 @@ export default {
 </script>
 
 <style lang="sass">
-  @import "./../../sass/_header.sass"
+  @import "@/sass/_header.sass"
 </style>
