+++ src/langs/ru.json
@@ -18,7 +18,7 @@
   "title-nick": "Данный псевдоним участника и его взнос будут указаны в общем списке участников",
   "forgot-password": "Забыли пароль?",
   "email-for-reset": "Введите эл.почту Вашего акаунта, мы вышлем ссылку для восстановления пароля",
-  "signed-in": "Вы авторизованы. Теперь у Вас есть доступ к проектам и Вы можете создать собственный проект.",
+  "signed-in": "Вы авторизованы. Теперь Вы можете создать собственный проект.",
   "agreement": "Согласен с условиями пользовательского соглашения",
   "close": "Закрыть",
   "view-project": "Открыть проект",
@@ -35,7 +35,9 @@
   "preliminary": "Предварительная смета-отчёт",
   "table-head":"<th>Ресурсы</th><th>Количество</th><th>Цена за 1 единицу</th><th>Сумма</th>",
   "sum": "Итого",
-  "example-contacts": "Если посетитель сайта решит принять участие в каком-то из проектов ему нужно отправить контакт удобного средства связи (Email, номер телефона, номер в мессенджере, ссылка на страницу в соц.сети или другие). Контакт станет доступен автору проекта в его личном кабинете, он свяжется с потенциальным участником для согласования формы участия и сохранит полученную сумму взноса и псевдоним в публичный список участников.",
+  "example-contacts": "Если посетитель сайта решит принять участие в проекте ему нужно отправить контакт удобного средства связи (Email, номер телефона, номер в мессенджере, ссылка на страницу в соц.сети или другие). Контакт станет доступен автору проекта в личном кабинете и он свяжется с потенциальным участником для согласования формы участия. Согласованную сумму взноса и псеводним автор сохранит в личном кабинете и они появятся в публичном списке участников.",
+  "example-contacts-author": "Если посетитель сайта решит принять участие в проекте ему нужно отправить контакт удобного средства связи (Email, номер телефона, номер в мессенджере, ссылка на страницу в соц.сети или другие). Контакт станет доступен Вам (автору проекта) в личном кабинете и Вы сможете связаться с потенциальным участником для согласования формы участия. Согласованную сумму взноса Вам нужно сохранить в личном кабинете. После этого сумма взноса и псеводним участника появятся в публичном списке участников.",
+  "example-contacts-user": "Если Вы решите принять участие в проекте отправьте контакт удобного средства связи в форме ниже(Email, номер телефона, номер в мессенджере, ссылка на страницу в соц.сети или другие). Контакт станет доступен автору проекта в его личном кабинете, он свяжется с Вами для согласования формы участия и сохранит согласованную сумму взноса и псевдоним в публичный список участников.",
   "author": "Организатор проекта свяжется с Вами для согласования формы и условий участия",
   "send": "Отправить",
   "open-donors": "Открыть список участников",
@@ -103,5 +105,6 @@
   "creator-photo": "Автор проекта может добавить после окончания реализации проекта (или отредактировать) финальные фотографии проекта",
   "user-final": "После окончания реализации проекта автор может опубликовать ниже финальную смету-отчет израсходованных ресурсов на реализацию проекта.",
   "user-photo": "После окончания реализации проекта автор может опубликовать ниже финальные фотографии реализованного проекта.",
-  "projects-enter": "Для доступа к проектам необходимо ввести ID номер проекта"
+  "projects-enter": "Для доступа к проектам необходимо ввести ID номер проекта",
+  "user-page-empty": "У Вас нет активных проектов."
 }
