+++ src/views/donorsPage/ProjectDonors.vue
@@ -66,5 +66,5 @@ export default {
 </script>
 
 <style lang="sass">
-  @import "./../../sass/_donors.sass"
+  @import "@/sass/_donors.sass"
 </style>
