+++ src/views/mainPage/NotFoundComponent.vue
@@ -1,0 +1,23 @@
+<template>
+  <div>
+    <h2>Page is not found</h2>
+  </div>
+</template>
+
+<script>
+  export default {
+    data() {
+      return {
+      }
+    },
+    computed: {
+    },
+    methods: {
+    }
+  };
+</script>
+
+<style lang="sass" scoped>
+  h2
+    color: green
+</style>
